+++ example/lib/src/my_app.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 import 'package:flutter/material.dart';
 import 'package:pubstar_io/pubstar_io.dart';
 import 'package:pubstar_io_example/src/controller.dart';
@@ -31,6 +33,31 @@ class _MyAppState extends State<MyApp> {
       child: SizedBox(width: double.infinity, height: 100),
     );
   }
+
+  final customConfig =
+      Platform.isAndroid
+          ? PubstarNativeCustomConfig(
+            layoutName: "pubstar_admob_native_big",
+            advertiserTextViewId: "ad_advertiser",
+            iconImageViewId: "ad_logo",
+            titleTextViewId: "ad_headline",
+            mediaContentViewGroupId: "ad_media",
+            bodyTextViewId: "ad_body",
+            callToActionButtonId: "ad_call_to_action",
+            loadingViewId: "pubstar_shimmer_native_big",
+            ctaColorHex: "#FFFFFF",
+          )
+          : PubstarNativeCustomConfig(
+            layoutName: "AppAdmobNativeCustom",
+            advertiserTextViewId: "1",
+            iconImageViewId: "2",
+            titleTextViewId: "3",
+            mediaContentViewGroupId: "4",
+            bodyTextViewId: "5",
+            callToActionButtonId: "6",
+            loadingViewId: "AppShimmerBanner",
+            ctaColorHex: "#FFFFFF",
+          );
 
   @override
   Widget build(BuildContext context) {
@@ -73,8 +100,8 @@ class _MyAppState extends State<MyApp> {
                   )
                   : buildPlaceholder(),
 
-              Text('Native AD'),
               const SizedBox(height: 10),
+              Text('Native AD'),
               _isAdSdkReady
                   ? ColoredBox(
                     color: Colors.grey.shade300,
@@ -102,36 +129,71 @@ class _MyAppState extends State<MyApp> {
                     ),
                   )
                   : buildPlaceholder(),
-              const SizedBox(height: 10),
 
-              // Text('Video AD'),
-              // const SizedBox(height: 10),
-              // _isAdSdkReady
-              //     ? ColoredBox(
-              //       color: Colors.grey.shade300,
-              //       child: SizedBox(
-              //         width: double.infinity,
-              //         height: 200,
-              //         child: PubstarVideoAdView(
-              //           adId: AdIdExample.videoId,
-              //           media:
-              //               'https://storage.googleapis.com/gvabox/media/samples/stock.mp4',
-              //           onError:
-              //               (error) => debugPrint(
-              //                 "FLUTTER - app: onError video, code: ${error.code} - rawCode: ${error.name}",
-              //               ),
-              //           onHide:
-              //               (reward) => debugPrint(
-              //                 "FLUTTER - app: onHide video  - type: ${reward?.type ?? ""} - amount: ${reward?.amount ?? 0}",
-              //               ),
-              //           onLoaded:
-              //               () => debugPrint("FLUTTER - app: onLoaded video"),
-              //           onShowed:
-              //               () => debugPrint("FLUTTER - app: onShowed video"),
-              //         ),
-              //       ),
-              //     )
-              //     : buildPlaceholder(),
+              const SizedBox(height: 10),
+              Text('Custom Native AD'),
+              _isAdSdkReady
+                  ? ColoredBox(
+                    color: Colors.grey.shade300,
+                    child: SizedBox(
+                      width: double.infinity,
+                      height: 220,
+                      child: PubstarAdView(
+                        adId: AdIdExample.nativeId,
+                        mode: PubstarAdViewMode.loadAndShow,
+                        type: PubstarAdType.native,
+                        size: PubstarAdSize.small,
+                        nativeCustomConfig: customConfig,
+                        onError:
+                            (error) => debugPrint(
+                              "FLUTTER - app: onError custom native, code: ${error.code} - rawCode: ${error.name}",
+                            ),
+                        onHide:
+                            (reward) => debugPrint(
+                              "FLUTTER - app: onHide custom native  - type: ${reward?.type ?? ""} - amount: ${reward?.amount ?? 0}",
+                            ),
+                        onLoaded:
+                            () => debugPrint(
+                              "FLUTTER - app: onLoaded custom native",
+                            ),
+                        onShowed:
+                            () => debugPrint(
+                              "FLUTTER - app: onShowed custom native",
+                            ),
+                      ),
+                    ),
+                  )
+                  : buildPlaceholder(),
+
+              Text('Video AD'),
+              const SizedBox(height: 10),
+              _isAdSdkReady
+                  ? ColoredBox(
+                    color: Colors.grey.shade300,
+                    child: SizedBox(
+                      width: double.infinity,
+                      height: 200,
+                      child: PubstarVideoAdView(
+                        adId: AdIdExample.videoId,
+                        media:
+                            'https://storage.googleapis.com/gvabox/media/samples/stock.mp4',
+                        type: PubStarVideoAdType.outStream,
+                        onError:
+                            (error) => debugPrint(
+                              "FLUTTER - app: video onError, code: ${error.code} - rawCode: ${error.name}",
+                            ),
+                        onHide:
+                            (reward) => debugPrint(
+                              "FLUTTER - app: video onHide  - type: ${reward?.type ?? ""} - amount: ${reward?.amount ?? 0}",
+                            ),
+                        onLoaded:
+                            () => debugPrint("FLUTTER - app: video onLoaded"),
+                        onShowed:
+                            () => debugPrint("FLUTTER - app: video onShowed"),
+                      ),
+                    ),
+                  )
+                  : buildPlaceholder(),
               const SizedBox(height: 10),
               Text('-------------------------'),