+++ ios/Classes/PubstarIoPlugin.swift
@@ -127,6 +127,7 @@ public class PubstarIoPlugin: NSObject, FlutterPlugin {
                 onLoaded: onAdLoadedCallback(callbackId: callbackId),
                 onError: onErrorCallback(callbackId: callbackId, message: "loadAd is failed.")
             )
+            result(nil)
         case "showAd":
             guard let adId = call.extractAdId(result: result) else {
                 return
@@ -139,6 +140,7 @@ public class PubstarIoPlugin: NSObject, FlutterPlugin {
                 onShowed: onAdShowedCallback(callbackId: callbackId),
                 onError: onErrorCallback(callbackId: callbackId, message: "showAd is failed.")
             )
+            result(nil)
         case "showAdWithViewId":
             guard let adId = call.extractAdId(result: result) else { return }
             guard let viewId = call.extractViewId(result: result) else { return }
@@ -162,6 +164,7 @@ public class PubstarIoPlugin: NSObject, FlutterPlugin {
                 onShowed: onAdShowedCallback(callbackId: callbackId),
                 onError: onErrorCallback(callbackId: callbackId, message: "showedAd is failed.")
             )
+            result(nil)
         case "loadAndShowAd":
             guard let adId = call.extractAdId(result: result) else { return }
             guard let callbackId = call.extractCallbackIdId(result: result) else { return }
@@ -174,6 +177,7 @@ public class PubstarIoPlugin: NSObject, FlutterPlugin {
                 onShowed: onAdShowedCallback(callbackId: callbackId),
                 onShowedError: onErrorCallback(callbackId: callbackId, message: "onShowed is failed.")
             )
+            result(nil)
         case "loadAndShowAdWithViewId":
             guard let adId = call.extractAdId(result: result) else { return }
             guard let viewId = call.extractViewId(result: result) else { return }
@@ -199,6 +203,7 @@ public class PubstarIoPlugin: NSObject, FlutterPlugin {
                 onShowed: onAdShowedCallback(callbackId: callbackId),
                 onShowedError: onErrorCallback(callbackId: callbackId, message: "onShowed is failed.")
             )
+            result(nil)
         case "loadAndShowBannerAd":
             guard let adId = call.extractAdId(result: result) else { return }
             guard let viewId = call.extractViewId(result: result) else { return }
@@ -226,11 +231,13 @@ public class PubstarIoPlugin: NSObject, FlutterPlugin {
                 onShowed: onAdShowedCallback(callbackId: callbackId),
                 onShowedError: onErrorCallback(callbackId: callbackId, message: "onShowed is failed.")
             )
+            result(nil)
         case "loadAndShowNativeAd":
             guard let adId = call.extractAdId(result: result) else { return }
             guard let viewId = call.extractViewId(result: result) else { return }
             let size = call.extractNativeSize(result: result)
             guard let callbackId = call.extractCallbackIdId(result: result) else { return }
+            let customConfig = call.extractCustomConfig()
             
             guard let adView = PubstarAdViewRegistry.shared.views[viewId] else {
                 result(
@@ -251,12 +258,15 @@ public class PubstarIoPlugin: NSObject, FlutterPlugin {
                 onLoaded: onAdLoadedCallback(callbackId: callbackId),
                 onHide: onAdHideCallback(callbackId: callbackId),
                 onShowed: onAdShowedCallback(callbackId: callbackId),
-                onShowedError: onErrorCallback(callbackId: callbackId, message: "onShowed is failed.")
+                onShowedError: onErrorCallback(callbackId: callbackId, message: "onShowed is failed."),
+                customConfig: customConfig
             )
+            result(nil)
         case "loadAndShowVideoAd":
             guard let adId = call.extractAdId(result: result) else { return }
             guard let viewId = call.extractViewId(result: result) else { return }
             guard let media = call.extractMedia(result: result) else { return }
+            guard let type = call.extractTypeVideo(result: result) else { return }
             guard let callbackId = call.extractCallbackIdId(result: result) else { return }
             
             guard let adView = PubstarAdViewRegistry.shared.views[viewId] else {
@@ -270,20 +280,18 @@ public class PubstarIoPlugin: NSObject, FlutterPlugin {
                 return
             }
             
-            guard let url = URL(string: media) else {
-                return
-            }
-            
             PubstarAdManagerWrapper.loadAndShowVideoAd(
                 adId: adId,
                 view: adView,
-                media: AVPlayer(url: url),
+                media: media,
+                type: type,
                 onLoaderError: onErrorCallback(callbackId: callbackId, message: "onLoaded is failed."),
                 onLoaded: onAdLoadedCallback(callbackId: callbackId),
                 onHide: onAdHideCallback(callbackId: callbackId),
                 onShowed: onAdShowedCallback(callbackId: callbackId),
                 onShowedError: onErrorCallback(callbackId: callbackId, message: "onShowed is failed.")
             )
+            result(nil)
 
         default:
           result(FlutterMethodNotImplemented)